Support for columns participating in multiple indices (Issue #258)
Set AttributeUsage.AllowMultiple to true for IndexAttribute and
UniqueAttribute.

Added methods to SQLiteConnection that can be used to figure out which
unique constraint(s) would be violated by an insert/update operation.

Added a static class to encapsulate the pragma calls used to get index
metadata. This can be expanded to include other pragmas as needed.