Treat uint32 and int64 as SQLite integers
SQLite doesn't distinguish between int types,
it gladly stores between 1 and 64 bits of data
in an integer column without much fuss.

https://www.sqlite.org/datatype3.html

However, when setting an auto-incrementing PK,
the column must be explicitly typed as "integer"
within the SQL itself.

This fix passes the test suite on Win32/sqlite.dll
and when using CSharp-SQLite when specifying
USE_CSHARP_SQLITE as a build argument.