Rollback transactions if the commit/release fails
These methods can fail with BUSY without terminating the transaction.
This caused problems with maintaining the transaction depth and also
put a burdern on the user to try committing again, or rolling back.

To simplify things, an automatic rollback as been added.
Now when transactions fail with BUSY, they also rollback and terminate.

Fixes #604