Properties/Fields, Multiple Primary Keys
### Change between Properties and Fields for Classes ###
We needed the ability to use Fields instead of Properties for classes. So we decided to add a mode `TableMapping.ColumnMemberInfoType` to `SQLiteConnection`.
If set to `PropertyInfo`, all properties are read which is the ***default*** behaviour.
When set to `FieldInfo`, all public fields are used.


### Add Multiple Primary Keys for a single Table ###
We needed multiple Primary Keys for a single Table, we added this in `CreateTable`.