Add FullTextTable class attribute &  FullTextIndexed property attribute for easy FTS indexing
Add simple attribute support for full text indexing of a subset of columns in table as well as support to change the default tokenizer behavior [for the table. Porting a change that I did in a copy of SQLite.cs for a UWP Japanese dictionary app ~9 years ago that I'm now trying to port to MAUI.  Added a bunch of unit tests to cover the functionality and confirmed the existing tests are all passing on Win11.

Example usage:

		[FullTextTable]
		public class Article
		{
			[PrimaryKey, AutoIncrement]
			public int Id { get; set; }

			[FullTextIndexed]
			public string Title { get; set; }

			[FullTextIndexed]
			public string Content { get; set; }
			public string Author { get; set; }
			public DateTime PublishDate { get; set; }
		}

				[FullTextTable("JapaneseText_FT", "unicode61")]
		public class JapaneseText
		{
			[PrimaryKey, AutoIncrement]
			public int Id { get; set; }

			[FullTextIndexed]
			public string Text { get; set; }

			public string Notes { get; set; }
		}